Primary Styling Done
Need to test with awkward browsers like edge. Most of the styling is done but I think I might have to make a few adjustments so that its a bit more responsive to larger form factor displays. Might add in a few more items to the page just for functionality